Rename controllers to lib
