+++ server/lib/file.explorer.js
