+++ server/routes/media.js
@@ -3,8 +3,8 @@ let fs = require('fs');
 
 let router = express.Router();
 const statusCodes = require('http-status-codes');
-let fileExplorer = require('../controllers').fileExplorer;
-let mediaProcessor = require('../controllers').media;
+let fileExplorer = require('../lib').fileExplorer;
+let mediaProcessor = require('../lib').media;
 
 /**
  * Streams a video file(.mp4).
